requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
python-jose
passlib
pydantic
email-validator
asyncpg
bcrypt
python-multipart